refactor: update query resolvers to use modular functions and improve code organization
